refactor(pagination): cleanup pagination to use same components
removed nested pagination component in favor of default values
