+++ app/hypercerts/[hypercertId]/page.tsx
@@ -9,8 +9,6 @@ import EvaluateButton from "@/components/hypercert/evaluate-button";
 import { CurrencyButtons } from "@/components/currency-buttons";
 import { ListForSaleButton } from "@/components/marketplace/list-for-sale-button";
 import ErrorState from "@/components/global/error-state";
-import { getHypercertAttestations } from "@/attestations/getHypercertAttestations";
-import EvaluationsList from "@/components/hypercert/evaluations-list";
 import HypercertListings from "@/components/marketplace/hypercert-listings";
 import HypercertEvaluations from "@/components/evaluations/hypercert-evaluations";
