+++ components/global/pagination/pagination-button.tsx
@@ -1,5 +1,5 @@
 import { ChevronLeft, ChevronRight } from "lucide-react";
-import { Button } from "./ui/button";
+import { Button } from "../../ui/button";
 import Link from "next/link";
 
 interface PaginationButtonProps {
